Revert "주문 연관테이블 (주문 제외) 및 포인트세부사항, 포인트 정책 테스틐 코드 작성 및 명세서 작성"

+++ src/main/java/com/nhnacademy/hexashoppingmallservice/controller/order/OrderStatusController.java
@@ -51,7 +51,15 @@ public class OrderStatusController {
     @DeleteMapping("/{orderStatusId}")
     public ResponseEntity<OrderStatus> deleteOrderStatus(@PathVariable Long orderStatusId, HttpServletRequest request) {
         jwtUtils.ensureAdmin(request);
-        orderStatusService.deleteOrderStatus(orderStatusId);
+        OrderStatus orderStatus = orderStatusService.getOrderStatus(orderStatusId);
+        if (Objects.isNull(orderStatus)) {
+            throw new OrderStatusNotFoundException(Long.toString(orderStatusId));
+        }
+        try {
+            orderStatusService.deleteOrderStatus(orderStatusId);
+        } catch (RuntimeException e) {
+            throw new SqlQueryExecuteFailException(e.getMessage());
+        }
         return ResponseEntity.noContent().build();
     }
 }
