+++ src/main/java/com/nhnacademy/hexashoppingmallservice/controller/order/WrappingPaperController.java
@@ -52,7 +52,15 @@ public class WrappingPaperController {
     @DeleteMapping("/{wrappingPaperId}")
     public ResponseEntity<WrappingPaper> deleteWrappingPaper(@PathVariable Long wrappingPaperId, HttpServletRequest request) {
         jwtUtils.ensureAdmin(request);
-        wrappingPaperService.deleteWrappingPaper(wrappingPaperId);
+        WrappingPaper wrappingPaper = wrappingPaperService.getWrappingPaper(wrappingPaperId);
+        if (Objects.isNull(wrappingPaper)) {
+            throw new WrappingPaperNotFoundException(Long.toString(wrappingPaperId));
+        }
+        try {
+            wrappingPaperService.deleteWrappingPaper(wrappingPaperId);
+        } catch (RuntimeException e) {
+            throw new SqlQueryExecuteFailException(e.getMessage());
+        }
         return ResponseEntity.noContent().build();
     }
 }
